moved get defect translations to defect repository
